Linux camera testing (#12)
* Rudimentary implementation, settings unfinished

* refactor: Close serial port in Mpc325's __del__ method if it is open

This commit modifies the Mpc325 class in mpc_hal.py to ensure that the serial port is closed in the __del__ method if it is open. This prevents potential resource leaks and improves the overall reliability of the code. By checking if the port is open before closing it, we avoid unnecessary operations and ensure that the port is only closed when necessary.

* linus testing updates